+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 
 
+## 2026-04-25
+- `09:07` -- **Updated** `apps/desktopProbe/src/server/jobScanner.ts`
+
 ## 2026-04-24
 - `18:17` -- **Updated** `apps/desktopProbe/src/pages/components/quietHoursSettings.tsx`
 - `18:14` -- **Updated** `apps/desktopProbe/src/pages/settings.tsx`